requirements.txt
requests
keyboard
colorama
Pillow